[feat] 안쓰이는 함수들 정리
