+++ src/renderer/components/topology/TopologySidePanel.tsx
@@ -6,14 +6,13 @@ import { getSchemaMap } from '@renderer/utils/storageAPI';
 import { selectCodeSelectedObjectInfo } from '@renderer/features/codeSliceInputSelectors';
 import { selectUiToggleSidePanel } from '@renderer/features/uiSliceInputSelectors';
 import { useAppDispatch, useAppSelector } from '@renderer/app/store';
-import { getObjectDataType } from '@renderer/types/terraform';
+import { getObjectDataType, getId } from '@renderer/types/terraform';
 import { TOP_NAVBAR_HEIGHT } from '../MainNavbar';
 import FormHeader from './state/form/Header';
 import FormTabs from './state/StateTabs';
 import preDefinedData from './state/form/utils/preDefinedData';
 import { setSelectedSourceSchema, ObjcectInfo } from '../../features/codeSlice';
 import { TOPOLOGY_TOOLBAR_HEIGHT } from './toolbar/TopologyToolbar';
-import { getId } from './state/form/utils/getResourceInfo';
 export const SIDEPANEL_WIDTH = 500;
 // 저장 버튼 누르면 redux objects에 content 덮어씌우기나이ㅓㄻ
 const TopologySidePanel = () => {
